Sign In Page Completed
Multiple new changes!
I created a context folder that will check if the user is signed in or not.
On start up, user will not be signed-in and will need to sign in with the button on the top-right.
We currently have a hard-coded username and password, it will be "Admin" and "passwd", they are case sensitive.
After user is signed-in, their username will follow them to the home screen.
Next we will use this name as a button to redirect the user to the user settings page, allowing them to log out.