Basic Search page created
Added a basic search page with a search icon/button at the top right of the app.
After pressing the search button, you will be redirected to the search page with basic search features such as year, genre, and name.
As of now, there is no search function until we add the comics' data and database.